requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==0.28.1
asyncio==3.4.3